fix: update gas report values and improve accuracy in ICommitReveal2 interface
This commit updates the gas report JSON file to reflect revised gas usage values for various operators in the `submitCv` scenarios, ensuring more accurate tracking of gas consumption. Additionally, the `estimateRequestPrice` function in the `ICommitReveal2` interface is modified to accept a `uint32` type for `callbackGasLimit`, enhancing type consistency and clarity in the function signature.

+++ src/ICommitReveal2.sol
@@ -11,9 +11,9 @@ pragma solidity ^0.8.30;
 interface ICommitReveal2 {
     function requestRandomNumber(uint32 callbackGasLimit) external payable returns (uint256);
 
-    function estimateRequestPrice(uint256 callbackGasLimit, uint256 gasPrice) external view returns (uint256);
+    function estimateRequestPrice(uint32 callbackGasLimit, uint256 gasPrice) external view returns (uint256);
 
-    function estimateRequestPrice(uint256 callbackGasLimit, uint256 gasPrice, uint256 numOfOperators)
+    function estimateRequestPrice(uint32 callbackGasLimit, uint256 gasPrice, uint256 numOfOperators)
         external
         view
         returns (uint256);
